perf: 暂存

+++ doro-orm/src/main/java/com/doro/orm/bean/CommentBean.java
@@ -54,8 +54,9 @@ public class CommentBean extends BaseBean {
 
     /**
      * 子评论数
+     * 是否应该将该字段放在这里，而不是单独创建一个表
      */
-    private Long comments;
+    private Integer comments;
 
     /**
      * 子评论列表
